+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [2.2.1] - 2020-12-02
 ### Changed
-- Disable travis cleanup
+- Fixed travis build
 
 ## [2.2.0] - 2020-12-02
 ### Changed